chore: Remove some builtins
Some of the builtinins rather belong into the Nova SDK.

First step is to identify to see what should go over. Then, remove it
here and then add it to the Nova SDK. Then implement it here.

For now, identify and delete here in order to get a slim, working
barebones Wandelscript.