fix(RPS-1311):  Improve error messages for `PlanTrajectoryErrors`
In the case of PlanTrajectoryErrors, omit the very long output error
lines in favor of a concise info about the error description.

This ALSO touches on the traceback, since it goes haywire in the CLI.
Simple solution is to not set the ProgramRunner._exc variable. The
output is so hyuge, thousands of lines of floating point numbers, nobody
in their right mind would inspect them and hope to see something
meaningful.

A somewhat technical but much lighter message is still conveyed to both
robot pad and stdout.

Also require nova 0.47.0 in order to get access to the new pretty string
method.